Изменил тип рейтинг на децимал, теперь он долежен сохраняться и загрузаться из бд. Есть баг с отключением игрока

+++ src/DoubleMaze/Game/Maze/RatingCalculator.cs
@@ -5,14 +5,14 @@ namespace DoubleMaze.Game.Maze
 {
     public class Rating
     {
-        public Rating(double value = 1000d)
+        public Rating(decimal value = 1000m)
         {
             Value = value;
         }
 
-        public double RoundValue => Round(Value, 1);
+        public decimal RoundValue => Round(Value, 1);
 
-        public double Value { get; private set; }
+        public decimal Value { get; private set; }
 
         public static void Update(Rating ra, Rating rb, bool raWin, bool rbWin)
         {
@@ -23,22 +23,22 @@ namespace DoubleMaze.Game.Maze
             rb.Value = Calc(rbValue, raValue, GetWinBalls(rbWin, raWin));
         }
 
-        private static double GetWinBalls(bool raWin, bool rbWin)
+        private static decimal GetWinBalls(bool raWin, bool rbWin)
         {
             if (raWin == rbWin)
-                return 0.5d;
+                return 0.5m;
 
             if (raWin)
-                return 1d;
+                return 1m;
 
-            return 0d;
+            return 0m;
         }
 
-        private static double Calc(double ra, double rb, double sa)
+        private static decimal Calc(decimal ra, decimal rb, decimal sa)
         {
-            double Ea = 1 / (1 + Pow(10, (rb - ra) / 400d));
+            decimal Ea = 1 / (1 + (decimal)Pow(10, (double)(rb - ra) / 400d));
 
-            const double K = 16d;
+            const decimal K = 16m;
             return ra + K * (sa - Ea);
         }
     }
